+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Read error that caused page crash in Dashboard when switching processes in settings
 - Fixed handling account switching with Wander
+- Fix display of total stake as ARIO instead of mARIO on gateway selector for redelegation
 
 ## [1.10.3] - 2025-02-25